requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
PyPDF2
pytest
httpx